+++ src/containerapp/azext_containerapp/_help.py
@@ -397,6 +397,11 @@ helps['containerapp service postgres'] = """
     short-summary: Commands to manage the postgres service for the Container Apps environment.
 """
 
+helps['containerapp service kafka'] = """
+    type: group
+    short-summary: Commands to manage the kafka service for the Container Apps environment.
+"""
+
 helps['containerapp service redis create'] = """
     type: command
     short-summary: Command to create the redis service.
@@ -407,6 +412,11 @@ helps['containerapp service postgres create'] = """
     short-summary: Command to create the postgres service.
 """
 
+helps['containerapp service kafka create'] = """
+    type: command
+    short-summary: Command to create the kafka service.
+"""
+
 helps['containerapp service redis delete'] = """
     type: command
     short-summary: Command to delete the redis service.
@@ -415,6 +425,11 @@ helps['containerapp service redis delete'] = """
 helps['containerapp service postgres delete'] = """
     type: command
     short-summary: Command to delete the postgres service.
+"""
+
+helps['containerapp service kafka delete'] = """
+    type: command
+    short-summary: Command to delete the kafka service.
 """
 
 helps['containerapp env update'] = """