+++ src/containerapp/HISTORY.rst
@@ -6,6 +6,10 @@ Release History
 Upcoming
 ++++++
 * 'az containerapp job execution show/list': improve table output format
+* 'az containerapp create/update': --yaml support properties for api-version 2023-04-01-preview (e.g. subPath, mountOptions)
+* 'az containerapp service': add support for creation and deletion of kafka
+* 'az containerapp create': --registry-server support registry with custom port
+* Add regex to fix validation for containerapp name
 * Add 'az containerapp ingress cors' for CORS support
 * 'az container app env create/update': support --enable-mtls parameter